requirements.txt
six
pandas
networkx==1.11
matplotlib
pydotplus
odfpy